debezium/dbz#2235 Map MySQL and MariaDB signed TINYINT columns to INT8
MySQL and MariaDB signed TINYINT is an 8-bit type (-128 to 127), yet
the ANTLR DDL parsers folded the TINYINT and INT1 keywords into
java.sql.Types.SMALLINT, so signed TINYINT columns were emitted with an
INT16 schema. This was the only integer mapping that did not use the
smallest Kafka Connect type capable of holding the source range;
Types.TINYINT already mapped to INT8 in the shared JDBC converters, but
no binlog column ever reached that path.

Register TINYINT and INT1 as Types.TINYINT in all three parsers so that
signed TINYINT columns take the core Types.TINYINT path and are emitted
as INT8 with Byte values. Unsigned handling keeps the established
widen-by-one-type rule via the type-name based overrides: a new
explicit rule emits TINYINT UNSIGNED (0 to 255) as INT16, alongside an
explicit MEDIUMINT UNSIGNED rule that documents the existing INT32
result now that the jdbcType no longer decides these cases.

Align convertTinyInt with the INT8 schema by producing Byte values
(previously it delegated to convertSmallInt and produced Short), and
extend the decimal-string default value handling to the tinyint
converter so defaults such as '10.01' keep working. Default values for
signed TINYINT columns are now emitted as Byte accordingly.

+++ debezium-connector-binlog/src/test/java/io/debezium/connector/binlog/BinlogDefaultValueIT.java
@@ -461,10 +461,10 @@ public abstract class BinlogDefaultValueIT<C extends SourceConnector> extends Ab
         Schema schemaC = record.valueSchema().fields().get(1).schema().fields().get(2).schema();
         Schema schemaD = record.valueSchema().fields().get(1).schema().fields().get(3).schema();
         Schema schemaE = record.valueSchema().fields().get(1).schema().fields().get(4).schema();
-        assertThat(schemaA.defaultValue()).isEqualTo((short) 0);
-        assertThat(schemaB.defaultValue()).isEqualTo((short) 1);
-        assertThat(schemaC.defaultValue()).isEqualTo((short) 1);
-        assertThat(schemaD.defaultValue()).isEqualTo((short) 1);
+        assertThat(schemaA.defaultValue()).isEqualTo((byte) 0);
+        assertThat(schemaB.defaultValue()).isEqualTo((byte) 1);
+        assertThat(schemaC.defaultValue()).isEqualTo((byte) 1);
+        assertThat(schemaD.defaultValue()).isEqualTo((byte) 1);
         assertThat(schemaE.defaultValue()).isEqualTo(null);
     }
 
@@ -490,13 +490,13 @@ public abstract class BinlogDefaultValueIT<C extends SourceConnector> extends Ab
         Schema schemaG = record.valueSchema().fields().get(1).schema().fields().get(6).schema();
         Schema schemaH = record.valueSchema().fields().get(1).schema().fields().get(7).schema();
 
-        assertThat(schemaA.defaultValue()).isEqualTo((short) 10);
+        assertThat(schemaA.defaultValue()).isEqualTo((byte) 10);
         assertThat(schemaB.defaultValue()).isEqualTo((short) 5);
         assertThat(schemaC.defaultValue()).isEqualTo(0);
         assertThat(schemaD.defaultValue()).isEqualTo(20L);
         assertThat(schemaE.defaultValue()).isEqualTo(null);
         assertEmptyFieldValue(record, "F");
-        assertThat(schemaG.defaultValue()).isEqualTo((short) 1);
+        assertThat(schemaG.defaultValue()).isEqualTo((byte) 1);
         assertThat(schemaH.defaultValue()).isEqualTo((int) 1);
     }
 
@@ -525,8 +525,8 @@ public abstract class BinlogDefaultValueIT<C extends SourceConnector> extends Ab
 
         Schema schemaA = record.valueSchema().fields().get(1).schema().fields().get(0).schema();
         Schema schemaB = record.valueSchema().fields().get(1).schema().fields().get(1).schema();
-        assertThat(schemaA.defaultValue()).isEqualTo((short) 1);
-        assertThat(schemaB.defaultValue()).isEqualTo((short) 0);
+        assertThat(schemaA.defaultValue()).isEqualTo((byte) 1);
+        assertThat(schemaB.defaultValue()).isEqualTo((byte) 0);
     }
 
     @Test
@@ -1023,7 +1023,7 @@ public abstract class BinlogDefaultValueIT<C extends SourceConnector> extends Ab
         Struct change = ((Struct) record.value()).getStruct("after");
         assertFieldDefaultValue(change, "C1", (short) 0);
         assertFieldDefaultValue(change, "C2", (short) 0);
-        assertFieldDefaultValue(change, "C3", (short) 0);
+        assertFieldDefaultValue(change, "C3", (byte) 0);
         assertFieldDefaultValue(change, "C4", "abc");
         assertFieldDefaultValue(change, "C5", "abc");
         assertFieldDefaultValue(change, "C6", "abc");
